requirements.txt
# ============================
# Core Framework
# ============================
langgraph>=0.2.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-anthropic>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0

# Tavily (web search tool for LangChain)
langchain-tavily>=0.1.0

# ============================
# Vector DB
# ============================
qdrant-client>=1.9.0

# ============================
# Embedding (SentenceTransformer)
# ============================
sentence-transformers>=3.0.0
torch>=2.0.0

# ============================
# Web Search
# ============================
tavily-python>=0.3.0

# ============================
# API
# ============================
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# ============================
# Data Processing
# ============================
pypdf2>=3.0.0
beautifulsoup4>=4.12.0
requests>=2.32.0
python-dotenv>=1.0.0
lxml>=5.0.0

# ============================
# Utilities
# ============================
tqdm>=4.66.0
loguru>=0.7.0
tenacity>=8.4.0
httpx>=0.27.0
